\name{adultsNeonates}
\alias{adultsNeonates}
\title{The adultsNeonates example data set of "stacked" data.}
\description{
   The adultsNeonates data is an example of stacked data from an ICS Flow Cytometry assay.
}
\usage{
   data(adultsNeonates) 
}

\value{data frame with 1 column of marker combination percentages and several cols of 'demographic'
data describing the percentage in a given row.}

\author{N. Hawkins, Fred Hutchinson Cancer Research Center, Seattle, WA}

\references{TR Kollmann, J Crabtree, A Rein-Weston, D Blimkie, F Thomai, X Wang, J Furlong, E Fortuno III,
A Hajjar, N Hawkins, S Self, and C Wilson, The neonatal innate immune system is not less responsive than 
the adult but responds differently, 2009, J Immunology, 183, 7150-7160}

\seealso{\code{\link{StackedData}} }

\examples{
# Load adultsNeoates data
data(adultsNeonates)
}
