\name{markerMatrix}
\alias{markerMatrix}
\title{An example of the markers data.}
\description{
   The markers data created to match the order of the marker combinations in each subset of interest
   in the adultsNeonates data.
}
\usage{
   data(markerMatrix) 
}

\value{matrix of 0's and 1's; rows = marker combinations, cols = marker names.}

\author{N. Hawkins, Fred Hutchinson Cancer Research Center, Seattle, WA}

\seealso{\code{\link{StackedData}} }

\examples{
# Load marker data
data(markerMatrix)
}
